add ascii, no animation, and cat =^_^=

+++ Program.cs
@@ -1,20 +1,49 @@
-﻿public class Bar 
+﻿using System.Diagnostics;
+
+public class Bar 
 {
     public int Width { get; set; }
     public int Ind { get; set; }
     public int Pct { get; set; }
-    public string Beg { get; } = "\uee00";
-    public string BegFilled { get; } = "\uee03";
-    public string Filler { get; } = "\uee04";
-    public string Space { get; } = "\uee01";
-    public string Head { get; } = "\uee01";
-    public string Fin { get; } = "\uee02"; 
-    public string FinFilled { get; } = "\uee05"; 
+    public int CatFrame { get; set; }
+    public long ElapsedMs { get; set; }
+    public string Beg { get; set; }
+    public string BegFilled { get; set; }
+    public string Filler { get; set; }
+    public string Space { get; set; }
+    public string Head { get; set; }
+    public string Fin { get; set; }
+    public string FinFilled { get; set; }
+}
+
+public class UnicodeTheme
+{
+    public const string Beg = "\uee00";
+    public const string BegFilled = "\uee03";
+    public const string Filler = "\uee04";
+    public const string Space = "\uee01";
+    public const string Head = "\uee01";
+    public const string Fin = "\uee02"; 
+    public const string FinFilled = "\uee05"; 
+}
+
+public class AsciiTheme
+{
+    public const string Beg = "[";
+    public const string BegFilled = "[";
+    public const string Filler = "=";
+    public const string Space = " ";
+    public const string Head = ">";
+    public const string Fin = "]"; 
+    public const string FinFilled = "]"; 
 }
 
 public class RmrfArgs 
 {
     public bool NoAnimation { get; set; }
+    public bool DryRun { get; set; }
+    public bool Ascii { get; set; }
+    public bool Cat { get; set; }
     public string Path { get; set; }
     public string[] Include { get; set; }
     public string[] Exclude { get; set; }
@@ -22,46 +51,68 @@ public class RmrfArgs
 
 public class Program 
 {
+    private static string[] _cats = [
+        "(  ^ｰ^)ﾆｬﾝ",
+        "(  ^ｰ^)ﾆｬﾝ",
+        "ﾆｬﾝ(^ｰ^ ) ",
+        "ﾆｬﾝ(^ｰ^*) ",
+        "み (>ω<*) ",
+    ];
+
     public static void Main(string[] args)
     {
-        Console.OutputEncoding = System.Text.Encoding.UTF8;
         if(!TryParse(args, out var rmrfArgs))
         {
             return;
         }
+        Console.OutputEncoding = rmrfArgs.Ascii 
+            ? System.Text.Encoding.ASCII             
+            : System.Text.Encoding.UTF8;
+
         Console.Write("Analyzing...");
         var flatDirs = FlattenTargets(GetTargets(rmrfArgs));
 
         var success = true;
         Exception e = null;
         using var sw = CreateStream();
-        var bar = CreateBar();
+        var bar = CreateBar(rmrfArgs);
         Console.CancelKeyPress += delegate {
+            if (rmrfArgs.NoAnimation)
+                return;
             Recover(sw);
         };
 
-        DrawInitial(sw, bar);
+        var stopwatch = new Stopwatch();
+        DrawInitial(sw, bar, rmrfArgs);
         var deleted = 0;
         var total = flatDirs.Count;
+        stopwatch.Start();
         foreach (var directory in flatDirs)
         {
-            success = DeleteDirectory(directory, out e);
+            success = DeleteDirectory(directory, rmrfArgs, out e);
             if (!success)
                 break;
             deleted++;
             var (nextPct, nextInd) = CountProgress(bar, deleted, total);
-            MoveDelta(sw, bar, nextInd, nextPct);
+            MoveDelta(sw, bar, rmrfArgs, nextInd, nextPct, stopwatch.ElapsedMilliseconds);
         }
         if (success)
-            MoveDelta(sw, bar, bar.Width - 1, 100);
-        DrawComplete(sw, success, e);
+            MoveDelta(sw, bar, rmrfArgs, bar.Width - 1, 100, stopwatch.ElapsedMilliseconds);
+        DrawComplete(sw, rmrfArgs, success, e);
     }
 
-    public static bool DeleteDirectory(string path, out Exception e)
+    private static bool DeleteDirectory(string path, RmrfArgs args, out Exception e)
     {
         e = null;
         try 
         {
+            if (args.DryRun)
+            {
+                var time = Directory.GetDirectories(path).Length + 
+                    Directory.GetFiles(path).Length;
+                Thread.Sleep(time / 4);
+                return true;
+            }
             foreach (var dir in Directory.GetDirectories(path))
             {
                 Directory.Delete(dir);
@@ -84,13 +135,13 @@ public class Program
         return true;
     }
     
-    public static (int, int) CountProgress(Bar bar, int current, int total)
+    private static (int, int) CountProgress(Bar bar, int current, int total)
     {
         var pct = (double)current / total;
         return ((int)(pct * 100), (int)(pct * (bar.Width - 1)));
     }
 
-    public static StreamWriter CreateStream()
+    private static StreamWriter CreateStream()
     {
         var sw = new StreamWriter(Console.OpenStandardOutput());
         sw.AutoFlush = true;
@@ -98,16 +149,39 @@ public class Program
         return sw;
     }
 
-    public static Bar CreateBar()
+    private static Bar CreateBar(RmrfArgs args)
     {
         var bar = new Bar
         {
             Width = 30
         };
+        SetupTheme(bar, args);
         return bar;
     }
 
-    public static List<string> FlattenTargets(List<string> targets)
+    private static void SetupTheme(Bar bar, RmrfArgs args)
+    {
+        if (args.Ascii)
+        {
+            bar.Beg = AsciiTheme.Beg;
+            bar.BegFilled = AsciiTheme.BegFilled;
+            bar.Filler = AsciiTheme.Filler;
+            bar.Space = AsciiTheme.Space;
+            bar.Head = AsciiTheme.Head;
+            bar.Fin = AsciiTheme.Fin; 
+            bar.FinFilled = AsciiTheme.FinFilled; 
+        } else {
+            bar.Beg = UnicodeTheme.Beg;
+            bar.BegFilled = UnicodeTheme.BegFilled;
+            bar.Filler = UnicodeTheme.Filler;
+            bar.Space = UnicodeTheme.Space;
+            bar.Head = UnicodeTheme.Head;
+            bar.Fin = UnicodeTheme.Fin; 
+            bar.FinFilled = UnicodeTheme.FinFilled; 
+        }
+    }
+
+    private static List<string> FlattenTargets(List<string> targets)
     {
         var flat = new List<string>(targets.Count);
         foreach (var target in targets)
@@ -117,7 +191,7 @@ public class Program
         return flat;
     }
 
-    public static void CollectInner(string path, List<string> flat)
+    private static void CollectInner(string path, List<string> flat)
     {
         var innerDirs = Directory.GetDirectories(path);
         foreach (var innerDir in innerDirs)
@@ -127,7 +201,7 @@ public class Program
         flat.Add(path);
     }
 
-    public static List<string> GetTargets(RmrfArgs args)
+    private static List<string> GetTargets(RmrfArgs args)
     {
         var res = new List<string>();
         var path = args.Path;
@@ -167,7 +241,7 @@ public class Program
         }
     }
 
-    public static bool TryParse(string[] args, out RmrfArgs res)
+    private static bool TryParse(string[] args, out RmrfArgs res)
     {
         int ind = 0;
         res = new RmrfArgs();
@@ -192,25 +266,58 @@ public class Program
                 res.Exclude = args[ind].Split(',');
                 ind++;
             }
+            else if (args[ind] == "-ascii") 
+            {
+                ind++;
+                res.Ascii = true;
+            }
+            else if (args[ind] == "-dry") 
+            {
+                ind++;
+                res.DryRun = true;
+            }
             else if (args[ind] == "-na") 
             {
                 ind++;
                 res.NoAnimation = true;
-            } else {
+            } 
+            else if (args[ind] == "-cat") 
+            {
+                ind++;
+                res.Cat = true;
+            } 
+            else 
+            {
+                Console.WriteLine("Invalid argument list");
+                Console.WriteLine();
                 Console.WriteLine("rmrf [args]");
                 Console.WriteLine("args list:");
-                Console.WriteLine("    -p:  path");
-                Console.WriteLine("    -i:  directories to delete");
-                Console.WriteLine("    -e:  exclude paths");
-                Console.WriteLine("    -na: no animation");
+                Console.WriteLine("    -p <folder1>:                path");
+                Console.WriteLine("    -i <folder1,folder2,...>:    directories to delete");
+                Console.WriteLine("    -e <folder1,folder2,...>:    exclude paths");
+                Console.WriteLine("    -na:                         no animation");
+                Console.WriteLine("    -dry:                        dry run");
+                Console.WriteLine("    -ascii:                      ascii theme");
+                Console.WriteLine("    -cat:                        add a cat");
                 return false;
             }
+        }
+        if (res.Ascii)
+        {
+            res.Cat = false;
         }
         return true;
     }
 
-    public static void DrawInitial(StreamWriter sw, Bar bar)
+    private static void DrawInitial(StreamWriter sw, Bar bar, RmrfArgs args)
     {
+        if (args.NoAnimation)
+        {
+            OutText(sw, "\r\n");
+            OutText(sw, "Start deleting");
+            OutText(sw, "\r\n");
+            return;
+        }
         OutEsc(sw, "?25l");
         OutEsc(sw, "1G");
         OutText(sw, bar.Beg);
@@ -220,10 +327,20 @@ public class Program
         }
         OutText(sw, bar.Fin);
         OutText(sw, "   0%");
+        if (args.Cat)
+        {
+            OutText(sw, "    ");
+            OutText(sw, _cats[0]);
+        }
     }
 
-    public static void DrawComplete(StreamWriter sw, bool success, Exception e = null)
+    private static void DrawComplete(StreamWriter sw, RmrfArgs args, bool success, Exception e = null)
     {
+        if (args.NoAnimation)
+        {
+            DrawCompleteNoAnimation(sw, success, e);
+            return;
+        }
         OutEsc(sw, "0m");
         OutText(sw, "\r\n");
         if (success)
@@ -243,16 +360,41 @@ public class Program
         OutEsc(sw, "?25h");
     }
 
-    public static void Recover(StreamWriter sw)
+    private static void DrawCompleteNoAnimation(StreamWriter sw, bool success, Exception e = null)
+    {
+        if (success)
+        {
+            OutText(sw, "Over");
+        } else {
+            OutText(sw, "Failed");
+            OutText(sw, "\r\n");
+            OutText(sw, e.ToString());
+        }
+    }
+
+    private static void Recover(StreamWriter sw)
     {
         OutEsc(sw, "0m");
         OutEsc(sw, "?25h");
     }
 
-    public static void MoveDelta(StreamWriter sw, Bar bar, int nextInd, int nextPct)
+    private static void MoveDelta(
+        StreamWriter sw, 
+        Bar bar, 
+        RmrfArgs args, 
+        int nextInd, 
+        int nextPct, 
+        long elapsed)
     {
+        if (args.NoAnimation)
+        {
+            MoveDeltaNoAnimation(sw, bar, args, nextInd, nextPct, elapsed);
+            return;
+        }
+        var catFrame = GetCatFrame(bar, args, elapsed);
         if (nextInd == bar.Ind &&
-            nextPct == bar.Pct)
+            nextPct == bar.Pct &&
+            catFrame == bar.CatFrame)
             return;
 
         if (nextInd != bar.Ind)
@@ -285,17 +427,56 @@ public class Program
             OutText(sw, pctStr);
         }
 
+        if (catFrame != bar.CatFrame)
+        {
+            OutEsc(sw, 4 + 4 + bar.Width + "G");
+            OutText(sw, "    ");
+            OutText(sw, _cats[catFrame]);
+            bar.ElapsedMs = elapsed;
+        }
+
         bar.Ind = nextInd;
         bar.Pct = nextPct;
+        bar.CatFrame = catFrame;
     }
 
-    public static void OutEsc(StreamWriter sw, string str)
+    private static void MoveDeltaNoAnimation(
+        StreamWriter sw, 
+        Bar bar, 
+        RmrfArgs args, 
+        int nextInd, 
+        int nextPct, 
+        long elapsed)
+    {
+        const int delay = 10;
+        if (nextPct - bar.Pct >= delay || (nextPct == 100 && bar.Pct != 100))
+        {
+            OutText(sw, $"Deleted {nextPct}%");
+            OutText(sw, "\r\n");
+            bar.Pct = nextPct;
+        }
+    }
+
+    private static int GetCatFrame(Bar bar, RmrfArgs args, long elapsed)
+    {
+        if (!args.Cat)
+            return 0;
+
+        const int delay = 700;
+
+        if (elapsed - bar.ElapsedMs > delay)
+            return (bar.CatFrame + 1) % _cats.Length;
+
+        return bar.CatFrame;
+    }
+
+    private static void OutEsc(StreamWriter sw, string str)
     {
         sw.Write("\x1b[");
         sw.Write(str);
     }
 
-    public static void OutText(StreamWriter sw, string str)
+    private static void OutText(StreamWriter sw, string str)
     {
         sw.Write(str);
     }
